+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -29,6 +29,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Append selected option value in dropdown in Settings if it doesn't exist (#3178)
 - Allow storing JSON data for options in Settings (#3179)
 - Added convenient method to handle Ajax Requests (#3180)
+- Allow executing further functionality on the Settings page (#3181)
 
 ### Fixed